# doc/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Explore the content of ARLAS collections"
  version: "V0.1.0"
  title: "ARLAS Exploration API"
  contact:
    name: "Gisaia"
    url: "http://www.gisaia.com/"
    email: "contact@gisaia.com"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/arlas"
...
schemes:
- "http"
paths:
  /collections/{collection}:
    get:
      tags:
      - "collections"
      summary: "Get a collection reference"
      description: "Get a collection reference in ARLAS"
      operationId: "get"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "collection"
        in: "path"
        description: "collection"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/CollectionReference"
        404:
          description: "Collection not found."
          schema:
            $ref: "#/definitions/ArlasError"
        500:
          description: "Arlas Server Error."
          schema:
            $ref: "#/definitions/ArlasError"
    put:
      tags:
      - "collections"
      summary: "Add a collection reference"
      description: "Add a collection reference in ARLAS"
      operationId: "put"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "collection"
        in: "path"
        description: "collection"
        required: true
        type: "string"
      - in: "body"
        name: "collectionParams"
        description: "collectionParams"
        required: true
        schema:
          $ref: "#/definitions/CollectionReferenceParameters"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/CollectionReference"
        400:
          description: "JSON parameter malformed."
          schema:
            $ref: "#/definitions/ArlasError"
        500:
          description: "Arlas Server Error."
          schema:
            $ref: "#/definitions/ArlasError"
    delete:
      tags:
      - "collections"
      summary: "Delete a collection reference"
      description: "Delete a collection reference in ARLAS"
      operationId: "delete"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "collection"
        in: "path"
        description: "collection"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ArlasSuccess"
        404:
          description: "Collection not found."
          schema:
            $ref: "#/definitions/ArlasError"
        500:
          description: "Arlas Server Error."
          schema:
            $ref: "#/definitions/ArlasError"
  /collections:
    get:
      tags:
      - "collections"
      summary: "Get all collection references"
      description: "Get all collection references in ARLAS"
      operationId: "getAll"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters: []
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CollectionReference"
        500:
          description: "Arlas Server Error."
          schema:
            $ref: "#/definitions/ArlasError"
...
  /explore/{collection}/_describe:
    get:
      tags:
      - "explore"
      summary: "Describe"
      description: "Describe the structure and the content of the given collection. "
      operationId: "describe"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "collection"
        in: "path"
        description: "collection"
        required: true
        type: "string"
      - name: "pretty"
        in: "query"
        description: "Pretty print"
        required: false
        type: "boolean"
        default: false
      - name: "human"
        in: "query"
        description: "Human readable print"
        required: false
        type: "boolean"
        default: false
      - name: "max-age-cache"
        in: "query"
        description: "max-age-cache"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Successful operation"
...
definitions:
  CollectionReference:
    type: "object"
    properties:
      collection_name:
        type: "string"
        description: "The collection name"
      params:
        description: "The collection parameters"
        $ref: "#/definitions/CollectionReferenceParameters"
    description: "The reference to ARLAS collection that embed elasticsearch index\
      \ description."
  CollectionReferenceParameters:
    type: "object"
    properties:
      index_name:
        type: "string"
        description: "The collection's index name"
      type_name:
        type: "string"
        description: "The collection's type name"
      id_path:
        type: "string"
        example: "id"
        description: "Path to the collection's id"
      geometry_path:
        type: "string"
        example: "geometry"
        description: "Path to the collection's geometry"
      centroid_path:
        type: "string"
        example: "centroid"
        description: "Path to the collection's centroid"
      timestamp_path:
        type: "string"
        example: "timestamp"
        description: "Path to the collection's timestamp"
    description: "The description of the elasticsearch index and the way ARLAS API\
      \ will serve it."
  ArlasError:
    type: "object"
    properties:
      status:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      error:
        type: "string"
  ArlasSuccess:
    type: "object"
    properties:
      status:
        type: "integer"
        format: "int32"
      message:
        type: "string"
...